# docs/openapi.yaml
openapi: 3.1.0
info:
  title: JoyfulGenius API
  version: 1.0.0
  description: |
    Complete REST API for the JoyfulGenius learning platform.

    ## Roles
    - **User**  — student. Registers via `/users/register`, browses & buys courses.
    - **Tutor** — course helper. Created by an admin. Grades submissions inside scoped subjects.
    - **Admin** — course owner. Created by the developer. Full write access to courses they own.
    - **Developer** — platform operator. Created via one-time `/developer/register`. Manages admin accounts. Does NOT touch course content.

    ## Auth
    - Access tokens are short-lived JWTs sent as `Authorization: Bearer <token>`.
    - Refresh tokens live in an httpOnly cookie (`refreshToken` / `adminRefreshToken`) scoped to the refresh endpoints.
    - The frontend stores access tokens in `localStorage`/`sessionStorage` keyed by role.

    ## Standard error envelope
    ```json
    { "message": "human-readable error", "code": "optional_machine_code" }
    ```
  contact:
    name: JoyfulGenius Support
    email: support@joyfulgenius.org
  license:
    name: Proprietary
servers:
  - url: http://localhost:4001/api
    description: Local development
  - url: https://api.joyfulgenius.org/api
    description: Production

# ─────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    userBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Student access token (`localStorage.token` on the frontend). Verified by `protect` middleware.
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Admin / developer / tutor access token (`localStorage.admin_token`). Verified by `verifyAdmin`.
        Additional role gates (`verifyDeveloper`, `requireCourseScope`) may narrow further.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refreshToken
      description: httpOnly cookie set by `/users/login`. Scoped to `/api/users/auth/refresh`.
    adminRefreshCookie:
      type: apiKey
      in: cookie
      name: adminRefreshToken
      description: httpOnly cookie set by admin/developer login. Scoped to `/api/admin/refresh` and `/api/developer/refresh`.

  # ── Reusable parameters ────────────────────────────────────────────
  parameters:
    ObjectIdParam:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId (24 hex chars)
      schema:
        type: string
        pattern: '^[0-9a-fA-F]{24}$'

  # ── Reusable responses ─────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not allowed (wrong role / scope / ownership)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        message: { type: string, example: "Something went wrong" }
        error:   { type: string, description: "Legacy alias — some endpoints use `error` instead of `message`" }
        redirect:
          type: string
          description: If the wrong login endpoint was hit, points at the correct one.
          example: "/developer/login"
      required: []
    ObjectId:
      type: string
      pattern: '^[0-9a-fA-F]{24}$'
      example: '68a22e67ba843698d32c427a'
    Timestamps:
      type: object
      properties:
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PaidCourseEntry:
      type: object
      properties:
        courseId:  { $ref: '#/components/schemas/ObjectId' }
        isPaid:    { type: boolean }
        username:  { type: string, nullable: true }
        serial:    { type: integer, nullable: true }
        paidAt:    { type: string, format: date-time, nullable: true }
        joinedAt:  { type: string, format: date-time, nullable: true }
        progress:
          type: object
          properties:
            completedLessons: { type: integer, example: 4 }
            totalLessons:     { type: integer, example: 12 }
        testResults:
          type: array
          items: { $ref: '#/components/schemas/ObjectId' }
        completedLectureIds:
          type: array
          items: { $ref: '#/components/schemas/ObjectId' }
    User:
      allOf:
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id:         { $ref: '#/components/schemas/ObjectId' }
            f_name:      { type: string, example: "Kalpak" }
            last_name:   { type: string, example: "Bhoir" }
            email:       { type: string, format: email }
            whatsappNo:  { type: string, example: "9876543210" }
            district:    { type: string, example: "Hingoli" }
            verified:    { type: boolean, example: true }
            enrolledCourses:
              type: array
              items: { $ref: '#/components/schemas/ObjectId' }
            paidCourses:
              type: array
              items: { $ref: '#/components/schemas/PaidCourseEntry' }
            tokenVersion: { type: integer, example: 0 }
    AdminSubjectScope:
      type: object
      description: One entry pairs a course with the tutor's subjects inside it. Empty subjectIds means "all subjects".
      properties:
        courseId:   { $ref: '#/components/schemas/ObjectId' }
        subjectIds:
          type: array
          items: { $ref: '#/components/schemas/ObjectId' }
    Admin:
      allOf:
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id:   { $ref: '#/components/schemas/ObjectId' }
            email: { type: string, format: email }
            name:  { type: string, example: "Kalpak" }
            role:
              type: string
              enum: [admin, developer, tutor]
            scopedCourseIds:
              type: array
              description: Tutor scope — course IDs the tutor may work on. Empty for admin/developer.
              items: { $ref: '#/components/schemas/ObjectId' }
            subjectScopes:
              type: array
              description: Tutor scope — per-course subject narrowing. Empty for admin/developer.
              items: { $ref: '#/components/schemas/AdminSubjectScope' }
            tokenVersion: { type: integer }
...
    LoginResponse:
      type: object
      properties:
        message:     { type: string }
        accessToken: { type: string, description: "Short-lived JWT (15 min)" }
        role:        { type: string }
        user:
          oneOf:
            - $ref: '#/components/schemas/User'
            - $ref: '#/components/schemas/Admin'
...
paths:
...
  /admin/login:
    post:
      tags: [Auth — Admin]
      summary: Admin / tutor login
      description: |
        Accepts admin + tutor accounts only. Developer accounts get 403 with a
        `redirect: "/developer/login"` hint so the UI can bounce them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Signed in
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /materials/{id}/stream:
    get:
      tags: [Study Materials]
      summary: Stream a material's binary payload
      description: 'Sets the appropriate `Content-Type` and (for PDFs) `Content-Disposition: inline`.'
      security: [{ userBearer: [] }]
      parameters: [ { $ref: '#/components/parameters/ObjectIdParam' } ]
      responses:
        '200':
          description: Binary stream
          content:
            application/octet-stream: {}
...